+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.1.1] - 2018-08-07
+### Added
+- Hooray, We have a sound.
+
 ## [0.1.0] - 2018-08-03
 ### Added
-- A simple snake game
+- A simple snake game.